+++ docs/api/mainpages/overview.h
@@ -15,28 +15,28 @@ Please be aware that Lomse is a work in progress, not having yet reached version
 
 ## License
 
-Lomse is distributed under the BSD 2-clause license, a permissive open source license to allow Lomse to be used in any projects, whether open source or proprietary. This license is a GPL compatible license, so you can use Lomse in GPL licensed projects (see @ref library-license).
+Lomse is distributed under the BSD 2-clause license, a permissive open source license to allow Lomse to be used in any projects, whether open source or proprietary. This license is a GPL compatible license, so you can use Lomse in GPL licensed projects (see @ref page-library-license).
 
-The code samples and code examples included in this API documentation are intended to show how the Lomse library can be used and so, intended to be copied. Therefore, they are released into the public domain (see @ref samples-license).
+The code samples and code examples included in this API documentation are intended to show how the Lomse library can be used and so, intended to be copied. Therefore, they are released into the public domain (see @ref page-code-license).
 
 
 ## Lomse usage
 
-If you are new to Lomse, @b please read the @subpage lomse-introduction and then read the topic or topics you are interested in:
+If you are new to Lomse, @b please read the @subpage page-introduction and then read the topic or topics you are interested in:
 
-- @subpage render-overview
-- @subpage print-api
-- @subpage document-edition-overview
-- @subpage sound-generation
-- @subpage tasks-page
+- @subpage page-render-overview
+- @subpage page-printing
+- @subpage page-edit-overview
+- @subpage page-sound-generation
+- @subpage page-tasks
 
 ## Other topics
 
-- @subpage file-formats
-- @subpage internal-model-overview
-- @subpage viewport-overview
-- @subpage lomse-callbacks
-- @subpage lomse-events
+- @subpage page-file-formats
+- @subpage page-internal-model
+- @subpage page-coordinates-viewport
+- @subpage page-callbacks
+- @subpage page-events
 
 
 */